# Add code mapper in frequency order