--- a/src/tests.rs
+++ b/src/tests.rs
@@ -1,2 +1,3 @@
 mod double_array_structure;
 mod iterator;
+mod mapper;