This PR added code mapping for CharwiseDaachorse.

The code mapping enables to represent most characters with small integers, resulting in faster construction, as follows.

```
unidic/build/daachorse/charwise                                                                          
                        time:   [1.7010 s 1.7037 s 1.7081 s]
                        change: [-37.270% -37.017% -36.795%] (p = 0.00 < 0.05)
                        Performance has improved.
```

There is no significant difference in matching time and memory efficiency.

```
unidic/find_overlapping/daachorse/charwise                                                                            
                        time:   [6.5671 ms 6.5730 ms 6.5790 ms]
                        change: [-0.3567% +0.4832% +1.1118%] (p = 0.24 > 0.05)
                        No change in performance detected.
```

```
== data/unidic/unidic ==
# old version
daachorse (charwise): 23361332 bytes, 22.279 MiB
# new version
daachorse (charwise): 23413260 bytes, 22.329 MiB
```